requirements.txt
gql[aiohttp]
typer[all]
python-dotenv
pydantic